refactor: delete not used files
