+++ src/service/api-actions.ts
@@ -2,10 +2,10 @@ import {createAsyncThunk} from '@reduxjs/toolkit';
 import {AppDispatch} from '../hooks/use-app-dispatch.ts';
 import {State} from '../hooks/use-app-selector.ts';
 import {AxiosInstance} from 'axios';
-import {ChosenOffer, Offer} from '../mocks/offers.ts';
+import {ChosenOffer, Offer} from '../store/offers-process/offers-process.slice.ts';
 import {APIPaths, MAX_REVIEWS_ON_PAGE, Paths} from '../const.ts';
 import {dropToken, saveToken} from './token.ts';
-import {Review} from '../mocks/reviews.ts';
+import {Review} from '../store/offers-process/offers-process.slice.ts';
 import {CommentWithOfferId} from '../components/commentary-send-form/comment-send-form.tsx';
 import {redirectToRoute} from '../store/action.ts';
 import {loadUserData, UserInfo} from '../store/user-process/user-process.slice.ts';
