+++ src/store/offers-process/offers-process.slice.ts
@@ -1,9 +1,62 @@
 import {createSlice} from '@reduxjs/toolkit';
 import {CITIES, MAX_REVIEWS_ON_PAGE, NameSpace} from '../../const.ts';
-import {ChosenOffer, Offer} from '../../mocks/offers.ts';
 import {shuffleNearby} from '../../utils.ts';
-import {Review} from '../../mocks/reviews.ts';
 import {allowedSortMethods, SortMethods} from '../../components/places-sorting-form/places-sorting-form.tsx';
+
+export type PlaceLocation = {
+    latitude: number;
+    longitude: number;
+    zoom: number;
+};
+
+type UserInfo = {
+    name: string;
+    avatarUrl: string;
+    isPro: boolean;
+};
+
+export type Review = {
+    id: string;
+    date: string;
+    user: UserInfo;
+    comment: string;
+    rating: number;
+};
+
+
+export type CityLocation = {
+    name: string;
+    location: PlaceLocation;
+};
+
+export type HostInfo = {
+    name: string;
+    avatarUrl: string;
+    isPro: boolean;
+}
+
+export type Offer = {
+    id: string;
+    title: string;
+    type: string;
+    price: number;
+    city: CityLocation;
+    location: PlaceLocation;
+    isFavorite: boolean;
+    isPremium: boolean;
+    rating: number;
+    previewImage: string;
+};
+
+export type ChosenOffer = Omit<Offer, 'previewImage'> & {
+    description: string;
+    bedrooms: number;
+    goods: string[];
+    host: HostInfo;
+    images: string[];
+    maxAdults: number;
+};
+
 
 export type City = typeof CITIES[number];
 
@@ -14,12 +67,12 @@ export type OfferInfo = {
 };
 
 type InitialOffersState = {
-  offersCity: City;
-  offers: Offer[];
-  savedOrderOffers: Offer[];
-  favoriteOffers: Offer[];
-  sortMethod: allowedSortMethods;
-  chosenOffer: OfferInfo;
+    offersCity: City;
+    offers: Offer[];
+    savedOrderOffers: Offer[];
+    favoriteOffers: Offer[];
+    sortMethod: allowedSortMethods;
+    chosenOffer: OfferInfo;
 }
 
 const initialOffersState: InitialOffersState = {
@@ -39,28 +92,28 @@ export const offersProcess = createSlice({
   name: NameSpace.Offers,
   initialState: initialOffersState,
   reducers: {
-    loadOffers: (state, action: {payload: Offer[]}) => {
+    loadOffers: (state, action: { payload: Offer[] }) => {
       state.offers = action.payload;
       state.savedOrderOffers = action.payload;
     },
-    loadNearbyOffers: (state, action: {payload: Offer[]}) => {
+    loadNearbyOffers: (state, action: { payload: Offer[] }) => {
       state.chosenOffer.nearbyOffers = shuffleNearby(action.payload);
     },
-    loadChosenOffer: (state, action: {payload: ChosenOffer}) => {
+    loadChosenOffer: (state, action: { payload: ChosenOffer }) => {
       state.chosenOffer.offerDetails = action.payload;
     },
-    loadOfferReviews: (state, action: {payload: Review[]}) => {
+    loadOfferReviews: (state, action: { payload: Review[] }) => {
       state.chosenOffer.offerReviews = action.payload.slice(-MAX_REVIEWS_ON_PAGE).reverse();
     },
-    loadFavoriteOffers: (state, action: {payload: Offer[]}) => {
+    loadFavoriteOffers: (state, action: { payload: Offer[] }) => {
       state.favoriteOffers = action.payload;
     },
-    changeOffersCity: (state, action: {payload: City}) => {
+    changeOffersCity: (state, action: { payload: City }) => {
       state.offersCity = action.payload;
     },
-    sortOffers: (state, action: {payload: allowedSortMethods}) => {
+    sortOffers: (state, action: { payload: allowedSortMethods }) => {
       state.sortMethod = action.payload;
-      switch(state.sortMethod){
+      switch (state.sortMethod) {
         case SortMethods.ByPopularity:
           state.offers = state.savedOrderOffers;
           break;
